refactor : applied the VAC pattern to Screens
